services/web/requirements.txt
alembic==1.4.2
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dateparser==0.7.6
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
humanize==2.6.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
loguru==0.5.1
Mako==1.1.3
MarkupSafe==1.1.1
maya==0.6.1
numpy==1.19.1
pandas==1.1.0
pathspec==0.8.0
pendulum==2.1.2
pyarrow==1.0.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
pytzdata==2020.1
regex==2020.7.14
requests==2.24.0
six==1.15.0
snaptime==0.2.4
SQLAlchemy==1.3.18
stravaio==0.0.9
toml==0.10.1
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.10
Werkzeug==1.0.1
WTForms==2.3.3
psycopg2-binary==2.8.4
gunicorn==20.0.4
